refactor: fix PR #74 review comments — decouple contracts from core layer
- contracts/audio.py: remove all core dependencies (SignalMeta, TopicModel,
  StreamSubscriber, Priority). Move AudioTransport to host layer.
- contracts/speech.py: extract ASR to contracts/asr.py, remove SpeechTopic
  and TopicModel dependency. Keep speech focused on TTS/output.
- contracts/asr.py: new file — pure ASR contract (ear), symmetric to TTS.
- topics/audio.py: new file — AudioRuntimeTopic and SpeechTopic as TopicModel
  subclasses in implementation layer (like channels/bridges).
- core/mindflow/audio_signal.py: new file — AudioSignal/AudioAction, moved
  from contracts to core layer alongside nuclei.
- core/mindflow/audio_nucleus.py: move AudioNucleusMeta class definition from
  manifests to core; manifests now only holds factory instance.
- core/mindflow/priority_mindflow.py: remove hardcoded AudioNucleus registration.
- host/speech/capture/audio_transport.py: new file — AudioTransport ABC lives
  in host layer because it couples to core types.
- stream_tts_speech.py & speech_service_provider.py: revert to clean dev
  state — remove TopicService injection and speaker name logic.
- Update all imports across listener, ptt_listener, volcengine_asr, and tests.

This corrects the dependency topology: contracts layer no longer depends on
core layer. Topics and signals live in implementation layer, ASR is separated
from speech, and manifests contain only instances — no class definitions.

via claude code
