+++ .moss_ws/src/MOSS/manifests/nuclei.py
@@ -14,9 +14,8 @@ from ghoshell_moss.core.blueprint.mindflow import (
     Nucleus,
     SignalMeta,
     InputSignal,
-    Priority,
 )
-from ghoshell_moss.core.mindflow.audio_nucleus import AudioNucleus
+from ghoshell_moss.core.mindflow.audio_nucleus import audio_nucleus_factory
 from ghoshell_container import IoCContainer
 
 
@@ -36,32 +35,4 @@ class ExampleNucleusMeta(NucleusMeta):
         raise NotImplementedError("Example stub — not intended for runtime use")
 
 
-class AudioNucleusMeta(NucleusMeta):
-    """音频感知核工厂 — 生产监听 audio 信号的 AudioNucleus。"""
-
-    def name(self) -> str:
-        return "audio_nucleus"
-
-    def description(self) -> str:
-        return "audio perception signal nucleus — aggregates audio signals from ASR/listener"
-
-    def signals(self) -> list[type[SignalMeta]]:
-        from ghoshell_moss.contracts.audio import AudioSignal
-        return [AudioSignal]
-
-    def factory(self, container: IoCContainer) -> Nucleus:
-        return AudioNucleus(
-            name="audio_nucleus",
-            description="audio perception signal nucleus",
-            target_signal="audio",
-            default_prompt="User spoke via voice input. Process the speech.",
-            suppress_seconds=0.5,
-            buffer_size=5,
-            min_priority=Priority.WARNING,
-            pulse_beat_interval=3.0,
-            logger=container.force_fetch_logger(),
-        )
-
-
 example_nucleus_factory = ExampleNucleusMeta()
-audio_nucleus_factory = AudioNucleusMeta()
