+++ src/ghoshell_moss/contracts/audio.py
@@ -2,37 +2,22 @@
 Audio capture contracts — shared abstractions for system audio input.
 
 Capture source → raw PCM → transport → consumers (ASR, waveform, AI perception).
-Transport isolates audio core from Matrix/Zenoh, keeping contracts layer dependency-free.
 """
-import logging
 from abc import ABC, abstractmethod
-from enum import Enum
-from typing import Callable, Protocol
 
 import numpy as np
 from pydantic import BaseModel, Field
 
 from ghoshell_moss.contracts.configs import ConfigType
-from ghoshell_moss.contracts.speech import SpeechTopic
-from ghoshell_moss.core.blueprint.mindflow import Priority, SignalMeta
-from ghoshell_moss.core.blueprint.session import StreamSubscriber
-from ghoshell_moss.core.concepts.topic import TOPIC_MODEL, TopicModel, TopicService, TopicWindow
 
 __all__ = [
-    "AudioAction",
-    "AudioCaptureConfig",
-    "AudioCaptureSource",
-    "AudioChunk",
     "AudioFrameMeta",
+    "AudioChunk",
+    "AudioCaptureConfig",
+    "AudioRuntimeInfo",
+    "AudioCaptureSource",
     "AudioPullLatest",
-    "AudioRuntimeReporter",
-    "AudioRuntimeTopic",
     "AudioSequentialConsumer",
-    "AudioSignal",
-    "AudioTransport",
-    "Preemptable",
-    "SpeechEventEmitter",
-    "SpeechEventReceiver",
 ]
 
 
@@ -69,13 +54,8 @@ class AudioCaptureConfig(ConfigType):
         return "audio_capture"
 
 
-class AudioRuntimeTopic(TopicModel):
-    """Audio capture runtime state broadcast via TopicWindow (max_size=1).
-
-    Replaces the old tmp_storage one-shot write with a continuously
-    updatable topic — consumers get heartbeat, running state, and stream
-    location without polling the filesystem.
-    """
+class AudioRuntimeInfo(BaseModel):
+    """Runtime discovery — where the stream lives and whether it's alive."""
 
     running: bool = False
     stream_key: str = ""
@@ -83,14 +63,6 @@ class AudioRuntimeTopic(TopicModel):
     device_explain: str = ""
     started_at: float = 0.0
     last_heartbeat: float = 0.0
-
-    @classmethod
-    def topic_type(cls) -> str:
-        return "audio/runtime"
-
-    @classmethod
-    def default_topic_name(cls) -> str:
-        return "audio/runtime"
 
 
 class AudioCaptureSource(ABC):
@@ -136,125 +108,3 @@ class AudioSequentialConsumer(ABC):
 
     @abstractmethod
     async def __anext__(self) -> AudioChunk: ...
-
-
-class AudioTransport(ABC):
-    """Transport abstraction isolating audio core from Matrix/Session/Zenoh.
-
-    Audio capture only needs: publish PCM, subscribe to PCM, process lock,
-    topic broadcast, and a logger. How those are implemented (Zenoh, FileLocker,
-    TopicService) is the adapter's concern — not audio core's.
-
-    Single coupling point: MatrixAudioTransport in host/speech/capture/.
-    """
-
-    # -- PCM stream --
-    @abstractmethod
-    def pub_pcm(self, chunk: bytes) -> None:
-        """Publish a raw PCM chunk to the audio stream."""
-        ...
-
-    @abstractmethod
-    def sub_pcm_callback(self, on_chunk: Callable[[bytes], None]) -> Callable[[], None]:
-        """Subscribe to PCM stream via callback. Returns a release handle."""
-        ...
-
-    @abstractmethod
-    def sub_pcm_stream(self, maxsize: int) -> StreamSubscriber:
-        """Subscribe to PCM stream as an async iterable."""
-        ...
-
-    # -- process lock --
-    @abstractmethod
-    def acquire_lock(self) -> bool:
-        """Acquire cross-process lock for exclusive device access."""
-        ...
-
-    @abstractmethod
-    def release_lock(self) -> None:
-        """Release the process lock."""
-        ...
-
-    # -- topic broadcast --
-    @abstractmethod
-    def pub_topic(self, topic: TopicModel) -> None:
-        """Publish a topic via the transport's topic service."""
-        ...
-
-    @abstractmethod
-    def topic_window(self, model: type[TOPIC_MODEL], max_size: int) -> TopicWindow[TOPIC_MODEL]:
-        """Create a bounded sliding window over a topic stream."""
-        ...
-
-    # -- logger --
-    @property
-    @abstractmethod
-    def logger(self) -> logging.Logger:
-        """Logger for audio capture diagnostics."""
-        ...
-
-
-# ── AudioSignal — mindflow integration ──────────────────────────
-
-
-class AudioAction(str, Enum):
-    SPEECH_STARTED = "speech_started"
-    SPEECH_FINAL = "speech_final"
-    WAKE_WORD = "wake_word"
-    AUDIO_ALERT = "audio_alert"
-
-
-class AudioSignal(SignalMeta):
-    """Audio perception signal — listener → mindflow attention preemption.
-
-    ASR emits SPEECH_FINAL when a completed sentence is ready. The signal
-    carries the SpeechTopic payload and challenges the Ghost's current
-    attention via mindflow.
-    """
-
-    action: AudioAction
-    speech_topic: SpeechTopic | None = None
-
-    @classmethod
-    def signal_name(cls) -> str:
-        return "audio"
-
-    @classmethod
-    def priority(cls) -> Priority:
-        return Priority.WARNING
-
-
-# ── Optional capability protocols ───────────────────────────────
-
-
-class Preemptable(Protocol):
-    """Component can be interrupted by attention preemption.
-
-    TTS/Speech/Player optionally implement this. When mindflow's attention
-    challenge returns preempt, attenuate() is called on the current action's
-    associated component. resume() is called when the preempting impulse
-    completes.
-    """
-
-    def attenuate(self) -> None: ...
-
-    def resume(self) -> None: ...
-
-
-class SpeechEventEmitter(Protocol):
-    """Component can broadcast SpeechTopic events. Listener/ASR implement."""
-
-    @property
-    def topic_service(self) -> TopicService: ...
-
-
-class SpeechEventReceiver(Protocol):
-    """Component can receive SpeechTopic events. Context/subtitle/memory implement."""
-
-    def on_speech_topic(self, topic: SpeechTopic) -> None: ...
-
-
-class AudioRuntimeReporter(Protocol):
-    """Component can report runtime state. Capture/Player implement."""
-
-    def runtime_info(self) -> AudioRuntimeTopic: ...
