+++ src/ghoshell_moss/core/mindflow/priority_mindflow.py
@@ -1,6 +1,5 @@
-from ghoshell_moss.core.blueprint.mindflow import Impulse, Reaction, Attention, Priority
+from ghoshell_moss.core.blueprint.mindflow import Impulse, Reaction, Attention
 from ghoshell_moss.core.mindflow.base_mindflow import AbsMindflow
-from ghoshell_moss.core.mindflow.buffer_nucleus import BufferNucleus
 from ghoshell_moss.core.mindflow.priority_attention import PriorityProtectionAttention
 from ghoshell_moss.core.mindflow.input_signal_nucleus import InputSignalNucleus
 from ghoshell_moss.contracts import LoggerItf
@@ -42,7 +41,6 @@ def new_default_mindflow(
 
     预配置:
       - InputSignalNucleus — IM 红点式 input signal 聚合
-      - AudioNucleus — 音频感知信号聚合 (target_signal="audio")
       - PriorityProtectionAttention — 纯优先级 + 固定保护期仲裁
 
     最小用法:
@@ -64,18 +62,6 @@ def new_default_mindflow(
     mindflow.with_nucleus(InputSignalNucleus(
         buffer_size=input_buffer_size,
         suppress_seconds=input_suppress_seconds,
-        logger=logger,
-    ))
-    # 音频感知信号聚合
-    mindflow.with_nucleus(BufferNucleus(
-        name="audio_nucleus",
-        description="audio perception signal nucleus",
-        target_signal="audio",
-        default_prompt="User spoke via voice input. Process the speech.",
-        suppress_seconds=0.5,
-        buffer_size=5,
-        min_priority=Priority.WARNING,
-        pulse_beat_interval=3.0,
         logger=logger,
     ))
     return mindflow
