+++ src/ghoshell_moss/host/providers/speech_service_provider.py
@@ -1,7 +1,5 @@
 from ghoshell_moss.contracts.speech import Speech, TTS, StreamAudioPlayer
 from ghoshell_moss.contracts.logger import LoggerItf
-from ghoshell_moss.core.blueprint.matrix import Matrix
-from ghoshell_moss.core.concepts.topic import TopicService
 from ghoshell_moss.core.speech import BaseTTSSpeech
 from ghoshell_container import IoCContainer, Provider, INSTANCE
 
@@ -17,14 +15,8 @@ class TTSSpeechServiceProvider(Provider[Speech]):
         logger = con.force_fetch(LoggerItf)
         player = con.force_fetch(StreamAudioPlayer)
         tts = con.force_fetch(TTS)
-        topic_service = con.force_fetch(TopicService)
-        matrix = con.force_fetch(Matrix)
-        ghost_name = matrix.ghost_name
-        speaker_name = ghost_name if ghost_name and ghost_name != 'None' else "Ghost"
         return BaseTTSSpeech(
             logger=logger,
             player=player,
             tts=tts,
-            topic_service=topic_service,
-            speaker_name=speaker_name,
         )
