+++ src/ghoshell_moss/host/speech/capture/miniaudio_capture.py
@@ -16,10 +16,10 @@ from ghoshell_moss.contracts.audio import (
     AudioChunk,
     AudioFrameMeta,
     AudioPullLatest,
-    AudioRuntimeTopic,
     AudioSequentialConsumer,
-    AudioTransport,
 )
+from ghoshell_moss.host.speech.capture.audio_transport import AudioTransport
+from ghoshell_moss.topics.audio import AudioRuntimeTopic
 from ghoshell_moss.core.blueprint.session import StreamSubscriber
 
 __all__ = [
