+++ src/ghoshell_moss/topics/__init__.py
@@ -1,0 +1,7 @@
+"""
+Topic models for MOSS runtime events.
+
+Topics live in the implementation layer (like channels and bridges),
+not in contracts. They depend on core.concepts.topic and are consumed
+by the TopicService at runtime.
+"""
