+++ tests/ghoshell_moss/speech/test_matrix_audio_transport.py
@@ -1,7 +1,7 @@
 """Unit tests for MatrixAudioTransport — verify delegation to Matrix/Session/Workspace."""
 from unittest.mock import MagicMock
 
-from ghoshell_moss.contracts.audio import AudioTransport
+from ghoshell_moss.host.speech.capture.audio_transport import AudioTransport
 from ghoshell_moss.host.speech.capture.matrix_audio_transport import MatrixAudioTransport
